Страничка на флатер.
Добавил аватарку из локал ассета
